// src/modules/todo/infrastructure/persistence/kysely/models.kysely.d.ts
import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
                           ? ColumnType<S, I | undefined, U>
                           : ColumnType<T, T | undefined, T>;
export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export type migrations = {
    id: Generated<number>;
    timestamp: string;
    name: string;
};
export type todo = {
    id: string;
    title: string;
    isCompleted: Generated<boolean>;
    createdAt: Generated<Timestamp>;
    updatedAt: Generated<Timestamp>;
};
export type DB = {
    migrations: migrations;
    todo: todo;
};